+++ e2e/e2etests/test_sui_withdraw_restricted_address.go
@@ -13,10 +13,13 @@ import (
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
 )
 
-// TestSuiWithdrawRestrictedAddress tests a withdrawal to a restricted address that reverts to a revert address
+// TestSuiWithdrawRestrictedAddress tests that a withdrawal to a restricted address reverts to a revert address
+// the test also add a case to check withdrawal to Sui invalid address immediately fail, we don't add a dedicated test as this is a small logic
 func TestSuiWithdrawRestrictedAddress(r *runner.E2ERunner, args []string) {
 	require.Len(r, args, 1)
 	amount := utils.ParseBigInt(r, args[0])
+
+	// Restricted address
 
 	// ARRANGE
 	// Given receiver, revert address
@@ -57,4 +60,18 @@ func TestSuiWithdrawRestrictedAddress(r *runner.E2ERunner, args []string) {
 	revertBalanceAfter, err := r.SUIZRC20.BalanceOf(&bind.CallOpts{}, revertAddress)
 	require.NoError(r, err)
 	require.EqualValues(r, new(big.Int).Add(revertBalanceBefore, amount), revertBalanceAfter)
+
+	// Invalid address format
+	tx, err = r.GatewayZEVM.Withdraw(
+		r.ZEVMAuth,
+		[]byte("0x25db16c3ca555f6702c07860503107bb73cce9f6c1d6df00464529db15d5a5abaa"),
+		amount,
+		r.SUIZRC20Addr,
+		gatewayzevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
+	)
+	if err != nil {
+		// if error is not directly returned, check that the tx failed
+		receipt := utils.MustWaitForTxReceipt(r.Ctx, r.ZEVMClient, tx, r.Logger, r.ReceiptTimeout)
+		utils.RequiredTxFailed(r, receipt)
+	}
 }